Пока что не работает округление, остальные баги вроде-бы исправил

+++ include/ft_printf.h
@@ -52,6 +52,6 @@ typedef struct			t_doubleToChar
 
 
 void	ft_print_int(int nbr);
-void	ft_longDouble(long double nbr, unsigned int accuracy);
+void	ft_longDouble(long double nbr, long long accuracy);
 
 #endif
